New API for fetching publications.
Could not fetch all the publications with SerpApi because it uses pagination. As a result, I could only retrieve a constant number of publications at a time.
This made it difficult to obtain the publications of authors who have relatively fewer works compared to others.
New API for this operation: scholarly. The commit contains a very basic workflow of retriving an author's publications and calculating their number