Script that writes publications & their cite.
The Scholarly API does not return the names of the authors who wrote the publication. This generated a new problem because the page needs to display an author's publications. Therefore, I used the Serpapi API again. All of an author's publications should be fetched, so the script calculates the number of pages that need to be fetched using the results from the Scholarly API, as it returns all the publications